fix: notfification window based on theme
