+++ apps/extension/src/components/activity-item/activity-links/index.tsx
@@ -3,14 +3,14 @@ import { useStore } from '@src/store'
 import { Message } from '@radixdlt/crypto'
 import { CopyIcon, ExternalLinkIcon } from '@radix-ui/react-icons'
 import { Text, Box, Flex, StyledLink } from 'ui/src/components/atoms'
-import { Action, Transaction } from '@src/services/types'
+import { Action, Transaction } from '@src/types'
 import { getShortAddress } from '@src/utils/string-utils'
 import Button from 'ui/src/components/button'
 import ButtonTipFeedback from 'ui/src/components/button-tip-feedback'
 import { ToolTip } from 'ui/src/components/tool-tip'
 import { copyTextToClipboard } from '@src/utils/copy-to-clipboard'
 import { ActivityType } from '@src/components/activity-type'
-import { EXPLORER_URL } from '@src/containers/wallet-panel/config'
+import { EXPLORER_URL } from '@src/config'
 import { TransactionMessage } from '../transaction-message'
 
 interface IProps {
