+++ apps/extension/src/containers/wallet-panel/accounts/send-token/send-token-review/index.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import BigNumber from 'bignumber.js'
 import { useQueryClient } from 'react-query'
 import { useLocation } from 'wouter'
-import { Token } from '@src/services/types'
+import { Token } from '@src/types'
 import { FinalizeTransaction, SubmitSignedTransaction } from '@src/services/radix/transactions'
 import { BuiltTransactionReadyToSign } from '@radixdlt/application'
 import { useImmer } from 'use-immer'
@@ -15,7 +15,7 @@ import { InfoStatBlock } from '@src/components/info-stat-block'
 import { SlippageBox } from '@src/components/slippage-box'
 import { Z3usSpinnerAnimation } from '@src/components/z3us-spinner-animation'
 import { HardwareWalletReconnect } from '@src/components/hardware-wallet-reconnect'
-import { EXPLORER_URL } from '@src/containers/wallet-panel/config'
+import { EXPLORER_URL } from '@src/config'
 import { Box, Text, Flex, StyledLink } from 'ui/src/components/atoms'
 import { AlertDialog, AlertDialogTrigger, AlertDialogContent } from 'ui/src/components/alert-dialog'
 import { SendReceiveHeader } from '../../send-receive-header'
