--- apps/extension/src/containers/wallet-panel/config.tsx
@@ -1,38 +1,0 @@
-import React from 'react'
-import { AccountsIcon, StakingIcon, SettingsIcon } from 'ui/src/components/icons'
-
-export const APP_HEIGHT = 600
-export const APP_WIDTH = 360
-export const PANEL_HEIGHT = '552px'
-export const SLIDE_PANEL_HEADER_HEIGHT = 30
-export const SLIDE_PANEL_HEIGHT = 140
-export const SLIDE_PANEL_EXPAND_HEIGHT = 459
-
-export const EXPLORER_URL = 'https://explorer.radixdlt.com/#'
-
-export const ACCOUNTS = 'accounts'
-export const NFT = 'nft'
-export const SWAP = 'swap'
-export const STAKING = 'staking'
-export const SETTINGS = 'settings'
-
-export const routesInfo = {
-	[ACCOUNTS]: {
-		id: ACCOUNTS,
-		name: 'Accounts',
-		icon: <AccountsIcon />,
-		href: 'account',
-	},
-	[STAKING]: {
-		id: STAKING,
-		name: 'Staking',
-		icon: <StakingIcon />,
-		href: 'staking',
-	},
-	[SETTINGS]: {
-		id: SETTINGS,
-		name: 'Settings',
-		icon: <SettingsIcon />,
-		href: 'settings',
-	},
-}
